Code reorganization & clean-up
- Move UserModelPermissionMixin under helevents.models
- Clean up redundant super calls
- Move organization_can_be_edited_by to events.utils
- Extract new util function from UserDataSourceAndOrganizationMixin.user_data_source_and_organization (get_user_data_source_and_organization_from_request)
