+++ events/utils.py
@@ -2,6 +2,7 @@ import collections
 import re
 import warnings
 from datetime import datetime, timedelta
+from functools import cache
 from typing import Optional
 
 import pytz
@@ -9,11 +10,13 @@ from dateutil.parser import parse as dateutil_parse
 from django.conf import settings
 from django.db import transaction
 from django.utils import timezone
+from django.utils.translation import gettext_lazy as _
 from django_orghierarchy.models import Organization
-from rest_framework.exceptions import ParseError
+from rest_framework.exceptions import ParseError, PermissionDenied
 
 from events.models import DataSource, Keyword, Place
 from events.sql import count_events_for_keywords, count_events_for_places
+from helevents.models import User
 
 
 def convert_to_camelcase(s):
@@ -241,3 +244,46 @@ def get_or_create_default_organization() -> Optional[Organization]:
         defaults={"name": "Muu", "data_source": data_source},
     )
     return organization
+
+
+def organization_can_be_edited_by(organization: Organization, user):
+    return organization in user.admin_organizations.all()
+
+
+@cache
+def get_user_data_source_and_organization_from_request(
+    request,
+) -> tuple[DataSource, Organization]:
+    """Get the data source and organization for the user from the request.
+
+    :param request: the request object
+    :return: a tuple containing the data source and organization
+    """
+    from events.auth import ApiKeyAuth
+
+    # api_key takes precedence over user
+    if isinstance(request.auth, ApiKeyAuth):
+        data_source = request.auth.get_authenticated_data_source()
+        publisher = data_source.owner
+        if not publisher:
+            raise PermissionDenied(_("Data source doesn't belong to any organization"))
+    else:
+        # objects *created* by api are marked coming from the system data source unless api_key is provided
+        # we must optionally create the system data source here, as the settings may have changed at any time
+        system_data_source_defaults = {
+            "user_editable_resources": True,
+            "user_editable_organizations": True,
+        }
+        data_source, created = DataSource.objects.get_or_create(
+            id=settings.SYSTEM_DATA_SOURCE_ID, defaults=system_data_source_defaults
+        )
+        # user organization is used unless api_key is provided
+        user = request.user
+        if isinstance(user, User):
+            publisher = user.get_default_organization()
+        else:
+            publisher = None
+        # no sense in doing the replacement check later, the authenticated publisher must be current to begin with
+        if publisher and publisher.replaced_by:
+            publisher = publisher.replaced_by
+    return data_source, publisher
